+++ app/__init__.py
@@ -3,6 +3,7 @@ from flask import Flask
 from app.config import Configuracion, validar_configuracion_critica
 from app.rutas_mantenedores import bp_mantenedores
 from app.rutas import bp_principal
+from app.rutas_ejecuciones import bp_ejecuciones
 from app.rutas_scripts import bp_scripts
 from app.rutas_tareas import bp_tareas
 from app.rutas_usuarios import bp_usuarios
@@ -23,4 +24,5 @@ def crear_app():
     app.register_blueprint(bp_mantenedores)
     app.register_blueprint(bp_tareas)
     app.register_blueprint(bp_scripts)
+    app.register_blueprint(bp_ejecuciones)
     return app